requirements.txt
llama-index==0.10.18
pymupdf
pinecone-client
transformers
torch
python-dotenv
virtualenv
openai